+++ modular_sdk/models/tenant_settings.py
@@ -1,9 +1,9 @@
 from pynamodb.attributes import UnicodeAttribute, MapAttribute
 
+from pynamodb.indexes import GlobalSecondaryIndex
+from modular_sdk.models.pynamongo.models import ModularBaseModel
+
 from modular_sdk.models.base_meta import BaseMeta, TABLES_PREFIX
-from modular_sdk.models.pynamodb_extension.base_role_access_model import \
-    BaseRoleAccessModel
-from modular_sdk.models.pynamodb_extension.base_model import BaseGSI
 from pynamodb.indexes import AllProjection
 
 
@@ -14,7 +14,7 @@ VALUE = 'v'
 MODULAR_TENANT_SETTINGS_TABLE_NAME = 'TenantSettings'
 
 
-class KeyTenantNameIndex(BaseGSI):
+class KeyTenantNameIndex(GlobalSecondaryIndex):
     class Meta(BaseMeta):
         index_name = f'{KEY}-index'
         read_capacity_units = 1
@@ -24,7 +24,7 @@ class KeyTenantNameIndex(BaseGSI):
     key = UnicodeAttribute(hash_key=True, attr_name=KEY)
 
 
-class TenantSettings(BaseRoleAccessModel):
+class TenantSettings(ModularBaseModel):
     class Meta(BaseMeta):
         table_name = f'{TABLES_PREFIX}{MODULAR_TENANT_SETTINGS_TABLE_NAME}'